Mid_term_project/requirements.txt
# Core app
streamlit
langchain==0.1.17
langchain-community==0.3.23
transformers==4.40.1
sentence-transformers==2.6.1
protobuf==3.20.3

# Vectorstore
chromadb>=0.4.15
faiss-cpu             # Optional: only needed if switching between FAISS and Chroma

# Optional extras
torch==2.2.2
openai
python-dotenv         # Optional for local dev




